backend/requirements.txt
Django>=4.0.0,<4.1.0
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
pyyaml
requests
django-cors-headers
pytz